Adds a sensible workaround to connect.py for ipv6
Issue raised stating ipv6 connectivity was failing
due to missing "[]" around the address. This patch
includes the socket module and attempts to validate
an address to be ipv6. If successful it will add the
needed "[]" around the address. If an exception is
raised then that means the address is not a valid
ipv6 so we assume it is a hostname or ipv4 and treat
it normally.

Fixes #244