PyVmomi: Do not add closed connections to connection pool
Currently, if the conn gets closed from the server, we still add it to the conn pool. In case of regular connections, a new socket is created and this does not cause a problem.
However, in case of ssl tunnelling, this causes a issue because the new socket created will not be ssl enabled and hence will fail the handshake. This is an issue because we are not using httplib's builtin tunneling feature like 'set_tunnel' since it is for Python 2.7 and above while PyVmomi still supports Python 2.6