Add support to pyVmomi for versions for sub-namespaces
The Add method for _BuildVersions will get only
the first component of the namespace and use that as
a key for the dictionary. However that's a problem for
a namespace like hbr which has 3 sub-namespaces:
hbr.agent, hbr.cascadeAgent and hbr.replica because
newestVersions will only save hbr.replica.version.version5
as HbrReplicaTypes is the first one that's imported
in pyVmomi/__init__.py.
This change adds the following behavior: if you add a
version 'x.y.z.version.version1', it will try
to add it as 'x' (for backwards compatibility) and 'x.y.z'.