SoapStubAdapter: handle if server don't respond properly
In case of the server not responding in time, httplib under pyVmomi
will throw a BadStatusLine with an empty string (more precisely "''"),
which is not helpful at all.

This patch works around it by retrying the query after waiting a bit,
until it gets a different exception, a valid response, or exhaust
the max retries.