[WIP] Fix Exception leaks
In python2/3 Exception leaks are possible due to
"traceback reference cycle problem". See
https://cosmicpercolator.com/2016/01/13/exception-leaks-in-python-2-and-3/

In SoapAdapter.py/InvokeMethod, if vim runtime error was thrown,
Exception leak occurs. This is due to the fact that when obj is thrown as
Exception (at SoapAdapter.py:1375), obj is referencing the current frame
through obj.__traceback__, and the current frame is referencing obj. Thus,
a circular reference is created. The exception thrown (obj) will never get
GC'ed, causing a Exception and resource leak.

Also in task.py/WaitForTask, if vim runtime error was thrown,
Exception leak occurs. The reason is the same -- frame holds reference to
err and err holds reference to frame.

The solution would be removing all references to obj when raising the
Exception. This is accomplished by the 'try...finally...' idiom. Also,
`deserializer` needs to be del'ed, as it also holds a references to obj.
(obj is deserializer.deser.result).

Signed-off-by: Bo Gan <ganb@vmware.com>