fix pvmomi the proxy support (#799)
- ensure the proxy configuration is passed to `__GetElementTree()`
- reuse @blacksponge fix for the  `SSLTunnelConnection` class
- add `test_ssl_tunnelwith_cert_file` test to validate that `cert_file`
  and `key_file` are passed as expected.

The patch adds a new `HTTPProxyConnection` class to handle the `CONNECT`
based proxies.
The patch does not touch the `SSLTunnelConnection` class. Both are
really difference, and I think it's a better approach.

To summarize, if the user set `sslProxyPath`, we use the original method,
if instead `httpProxyHost` is set, we use the new class. Both should not
be used at the same time.

closes: #620
closes: #567
closes: #198