Mark non-remote ManagedObject and DataObject methods
"'Private' instance variables that cannot be accessed except from inside an object don’t exist in Python. However, there is a convention that is followed by most Python code: a name prefixed with an underscore (e.g. _spam) should be treated as a non-public part of the API (whether it is a function, a method or a data member). It should be considered an implementation detail and subject to change without notice." Ref. https://docs.python.org/3/tutorial/classes.html#private-variables

This change adds comment before the non-remote methods available for ManagedObject and DataObject that explains that those methods are not private but this is the way non-remote methods are marked in the VMODL1 context.