Support the import of old vsanMgmtObjects.py with new pyVmomi
We have moved the VSAN Python SDK into pyVmomi. But this creates a problem with existing customer deployments of the SDK. If they upgrade pyVmomi without modifying their code, when they import vsanmgmtObjects.py, it will try to register types which have already been registered, which will raise an exceotion and break their scripts.
To address the issue, this change adds to pyVmomi the ability to mark type definintions as "frozen" in bulk by turning on and off a flag which affects subsequent definitions. The flag is checked by the CreateDataType, CreateManagedType and CreateEnumType functions. Then the new functionality is used in _typeinfos.py to mark all vsan types as frozen, which eliminates the crash and ensures that the VSAN type definitions in pyVmomi will have precedence.