Add vSAN SDK pieces
The vSAN SDK provides two pieces of executable code:

 - vSAN health type info
 - API utils

Legacy scripts explicitly import the vsanmgmtObjects module, so we have to supply it in pyVmomi to avoid breakages. The module in pyVmomi is naturally empty.

The utils are also imported by scripts using the vSAN SDK, so with this change, these are also packaged in pyVmomi.